requirements.txt
asgiref==3.8.1
attrs==24.3.0
black==24.10.0
click==8.1.7
colorama==0.4.6
Django==4.1
django-debug-toolbar==3.2.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
load-dotenv==0.1.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pep8-naming==0.13.2
pillow==11.0.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.22.3
sqlparse==0.5.3
tzdata==2024.2
uritemplate==4.1.1
